chore: remove kotlin-based API server
I need to start smaller.

While I have a lot of ideas of what I want to do, there's also a point
in getting something at all out the door. Also, my research into
GraphQL and SQL under kotlin has indicated that there's a lot of
boilerplate involved to get these things working. Rust, on the other
hand, seems to have a more straightforward way of dealing with these
and should perform better.

The idea is therefore to pivot to a slim, rust-based, server with a
simpler data model and get this server up and running sooner rather
than later.